Implemented LinksServices
