+++ HW8/Evgeniy_Nikulin/MultithreadApp.Tests/MultithreadApp.Tests.csproj
@@ -64,7 +64,7 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="ApplicationTests.cs" />
+    <Compile Include="LinksServicesTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
