+++ HW8/Evgeniy_Nikulin/MultithreadApp/AppRegistry.cs
@@ -10,6 +10,7 @@
         public AppRegistry()
         {
             this.For<IApplication>().Use<Application>();
+            this.For<ILinksServices>().Use<LinksServices>();
             this.For<IDataBase>().Use<DataBaseUW>();
             this.For<IHttpProvider>().Use<WikiHttpProvider>();
             this.For<IFileProvider>().Use<FileProvider>();
